+++ src/my_app/lib/theming/dimens.dart
@@ -1,0 +1,12 @@
+// Flutter imports:
+import 'package:flutter/material.dart';
+
+const defaultPadding = 16.0;
+const defaultHorizontalMargin = 16.0;
+const cardCornerRadius = 24.0;
+const buttonCornerRadius = 12.0;
+const inputCornerRadius = 12.0;
+
+// Bottom Navigation Bar
+const bottomNavBarHeight = kBottomNavigationBarHeight;
+const bottomNavBarCornerRadius = 10.0;